exctracting User into a separate component
