+++ client/src/components/FriendsDashboard/SearchArea.js
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import styled from 'styled-components';
 
 export default function SearchArea ({ setQuery, query }) {
