+++ client/src/components/FriendsDashboard/User.js
@@ -1,0 +1,37 @@
+import styled from 'styled-components';
+
+import ListItem from '@material-ui/core/ListItem';
+import Avatar from '@material-ui/core/Avatar';
+import Divider from '@material-ui/core/Divider';
+import Button from '@material-ui/core/Button';
+
+export default function Suggestions ({ user }) {
+    return (
+        <>
+            <ListItem button>
+                <Avatar>{user.displayName[0]}</Avatar>
+                <Name>{user.displayName}</Name>
+                <SyledButton variant='outlined'>
+                    Follow
+                </SyledButton>
+            </ListItem>
+            <Divider />
+        </>
+    )
+}
+
+const Name = styled.div`
+    display: flex;
+    margin-left: 1em;
+    font-size: 0.9em;
+    font-weight: 600;
+`;
+
+const SyledButton = styled(Button)`
+    text-transform: uppercase;
+    width: 100px;
+    height: 25px;
+    background-color: lightgreen;
+    position: absolute;
+    right: 16px;
+`;
